enable player arrow on minimap per default
It just makes sense to have this.
This should reduce the number of questions asked on the Discord how to enable the arrow.

It should produce a red arrow in a good, sensible size. Red is a good color because it contrasts well with the usual background color (green) but also with stoney grey, snowy white etc. Even in the nether this shade of red seems to work quite well.